Move all D structures to d-tree.h
