+++ gcc/d/d-builtins.cc
@@ -35,6 +35,7 @@
 #include "stringpool.h"
 #include "stor-layout.h"
 
+#include "d-tree.h"
 #include "d-lang.h"
 #include "d-codegen.h"
 #include "d-objfile.h"
