+++ gcc/d/d-convert.cc
@@ -27,128 +27,7 @@
 #include "convert.h"
 #include "stor-layout.h"
 
-#include "d-lang.h"
-
-// Creates an expression whose value is that of EXPR, converted to type TYPE.
-// This function implements all reasonable scalar conversions.
-
-tree
-convert(tree type, tree expr)
-{
-  tree e = expr;
-  tree_code code = TREE_CODE(type);
-
-  if (type == error_mark_node
-      || expr == error_mark_node
-      || TREE_TYPE(expr) == error_mark_node)
-    return error_mark_node;
-
-  const char *invalid_conv_diag
-    = targetm.invalid_conversion(TREE_TYPE(expr), type);
-
-  if (invalid_conv_diag)
-    {
-      error("%s", invalid_conv_diag);
-      return error_mark_node;
-    }
-
-  if (type == TREE_TYPE(expr))
-    return expr;
-
-  if (TREE_CODE(type) == ARRAY_TYPE
-      && TREE_CODE(TREE_TYPE(expr)) == ARRAY_TYPE
-      && TYPE_DOMAIN(type) == TYPE_DOMAIN(TREE_TYPE(expr)))
-    return expr;
-
-  tree ret = targetm.convert_to_type(type, expr);
-  if (ret)
-    return ret;
-
-  STRIP_TYPE_NOPS(e);
-  tree etype = TREE_TYPE(e);
-
-  if (TYPE_MAIN_VARIANT(type) == TYPE_MAIN_VARIANT(TREE_TYPE(expr)))
-    return fold_convert(type, expr);
-  if (TREE_CODE(TREE_TYPE(expr)) == ERROR_MARK)
-    return error_mark_node;
-  if (TREE_CODE(TREE_TYPE(expr)) == VOID_TYPE)
-    {
-      error("void value not ignored as it ought to be");
-      return error_mark_node;
-    }
-
-  switch (code)
-    {
-    case VOID_TYPE:
-      return fold_convert(type, e);
-
-    case INTEGER_TYPE:
-    case ENUMERAL_TYPE:
-      if (TREE_CODE(etype) == POINTER_TYPE
-	  || TREE_CODE(etype) == REFERENCE_TYPE)
-	{
-	  if (integer_zerop(e))
-	    return build_int_cst(type, 0);
-
-	  // Convert to an unsigned integer of the correct width first, and
-	  // from there widen/truncate to the required type.
-	  tree utype = lang_hooks.types.type_for_size(TYPE_PRECISION(etype), 1);
-	  ret = fold_build1(CONVERT_EXPR, utype, e);
-	  return fold_convert(type, ret);
-	}
-
-      ret = convert_to_integer(type, e);
-      goto maybe_fold;
-
-    case BOOLEAN_TYPE:
-      return fold_convert(type, d_truthvalue_conversion(expr));
-
-    case POINTER_TYPE:
-    case REFERENCE_TYPE:
-      ret = convert_to_pointer(type, e);
-      goto maybe_fold;
-
-    case REAL_TYPE:
-      if (TREE_CODE (etype) == COMPLEX_TYPE && D_TYPE_IMAGINARY_FLOAT (type))
-	e = build1(IMAGPART_EXPR, TREE_TYPE (etype), e);
-
-      ret = convert_to_real(type, e);
-      goto maybe_fold;
-
-    case COMPLEX_TYPE:
-      if (TREE_CODE (etype) == REAL_TYPE && D_TYPE_IMAGINARY_FLOAT (etype))
-	ret = build2(COMPLEX_EXPR, type,
-		     build_zero_cst(TREE_TYPE (type)),
-		     convert(TREE_TYPE (type), expr));
-      else
-	ret = convert_to_complex(type, e);
-      goto maybe_fold;
-
-    case VECTOR_TYPE:
-      ret = convert_to_vector(type, e);
-      goto maybe_fold;
-
-    case RECORD_TYPE:
-    case UNION_TYPE:
-      if (lang_hooks.types_compatible_p(type, TREE_TYPE(expr)))
-	{
-	  ret = build1(VIEW_CONVERT_EXPR, type, expr);
-	  goto maybe_fold;
-	}
-      break;
-
-    default:
-      break;
-
-    maybe_fold:
-      if (!TREE_CONSTANT(ret))
-	ret = fold(ret);
-      return ret;
-    }
-
-  error("conversion to non-scalar type requested");
-  return error_mark_node;
-}
+#include "d-tree.h"
 
 
 // Build CODE expression with operands OP0 and OP1.
@@ -378,5 +257,127 @@ d_truthvalue_conversion (tree expr)
 
       return d_build_truthvalue_op (NE_EXPR, expr, integer_zero_node);
     }
+}
+
+
+// Creates an expression whose value is that of EXPR, converted to type TYPE.
+// This function implements all reasonable scalar conversions.
+
+tree
+convert(tree type, tree expr)
+{
+  tree e = expr;
+  tree_code code = TREE_CODE(type);
+
+  if (type == error_mark_node
+      || expr == error_mark_node
+      || TREE_TYPE(expr) == error_mark_node)
+    return error_mark_node;
+
+  const char *invalid_conv_diag
+    = targetm.invalid_conversion(TREE_TYPE(expr), type);
+
+  if (invalid_conv_diag)
+    {
+      error("%s", invalid_conv_diag);
+      return error_mark_node;
+    }
+
+  if (type == TREE_TYPE(expr))
+    return expr;
+
+  if (TREE_CODE(type) == ARRAY_TYPE
+      && TREE_CODE(TREE_TYPE(expr)) == ARRAY_TYPE
+      && TYPE_DOMAIN(type) == TYPE_DOMAIN(TREE_TYPE(expr)))
+    return expr;
+
+  tree ret = targetm.convert_to_type(type, expr);
+  if (ret)
+    return ret;
+
+  STRIP_TYPE_NOPS(e);
+  tree etype = TREE_TYPE(e);
+
+  if (TYPE_MAIN_VARIANT(type) == TYPE_MAIN_VARIANT(TREE_TYPE(expr)))
+    return fold_convert(type, expr);
+  if (TREE_CODE(TREE_TYPE(expr)) == ERROR_MARK)
+    return error_mark_node;
+  if (TREE_CODE(TREE_TYPE(expr)) == VOID_TYPE)
+    {
+      error("void value not ignored as it ought to be");
+      return error_mark_node;
+    }
+
+  switch (code)
+    {
+    case VOID_TYPE:
+      return fold_convert(type, e);
+
+    case INTEGER_TYPE:
+    case ENUMERAL_TYPE:
+      if (TREE_CODE(etype) == POINTER_TYPE
+	  || TREE_CODE(etype) == REFERENCE_TYPE)
+	{
+	  if (integer_zerop(e))
+	    return build_int_cst(type, 0);
+
+	  // Convert to an unsigned integer of the correct width first, and
+	  // from there widen/truncate to the required type.
+	  tree utype = lang_hooks.types.type_for_size(TYPE_PRECISION(etype), 1);
+	  ret = fold_build1(CONVERT_EXPR, utype, e);
+	  return fold_convert(type, ret);
+	}
+
+      ret = convert_to_integer(type, e);
+      goto maybe_fold;
+
+    case BOOLEAN_TYPE:
+      return fold_convert(type, d_truthvalue_conversion(expr));
+
+    case POINTER_TYPE:
+    case REFERENCE_TYPE:
+      ret = convert_to_pointer(type, e);
+      goto maybe_fold;
+
+    case REAL_TYPE:
+      if (TREE_CODE (etype) == COMPLEX_TYPE && D_TYPE_IMAGINARY_FLOAT (type))
+	e = build1(IMAGPART_EXPR, TREE_TYPE (etype), e);
+
+      ret = convert_to_real(type, e);
+      goto maybe_fold;
+
+    case COMPLEX_TYPE:
+      if (TREE_CODE (etype) == REAL_TYPE && D_TYPE_IMAGINARY_FLOAT (etype))
+	ret = build2(COMPLEX_EXPR, type,
+		     build_zero_cst(TREE_TYPE (type)),
+		     convert(TREE_TYPE (type), expr));
+      else
+	ret = convert_to_complex(type, e);
+      goto maybe_fold;
+
+    case VECTOR_TYPE:
+      ret = convert_to_vector(type, e);
+      goto maybe_fold;
+
+    case RECORD_TYPE:
+    case UNION_TYPE:
+      if (lang_hooks.types_compatible_p(type, TREE_TYPE(expr)))
+	{
+	  ret = build1(VIEW_CONVERT_EXPR, type, expr);
+	  goto maybe_fold;
+	}
+      break;
+
+    default:
+      break;
+
+    maybe_fold:
+      if (!TREE_CONSTANT(ret))
+	ret = fold(ret);
+      return ret;
+    }
+
+  error("conversion to non-scalar type requested");
+  return error_mark_node;
 }
