+++ gcc/d/d-decls.cc
@@ -37,6 +37,7 @@
 #include "stringpool.h"
 #include "stor-layout.h"
 
+#include "d-tree.h"
 #include "d-lang.h"
 #include "d-codegen.h"
 #include "d-objfile.h"
