+++ gcc/d/d-elem.cc
@@ -32,6 +32,7 @@
 #include "function.h"
 #include "stor-layout.h"
 
+#include "d-tree.h"
 #include "d-lang.h"
 #include "d-codegen.h"
 #include "d-objfile.h"
