+++ gcc/d/d-glue.cc
@@ -31,6 +31,7 @@
 #include "fold-const.h"
 #include "diagnostic.h"
 
+#include "d-tree.h"
 #include "d-lang.h"
 #include "d-objfile.h"
 #include "d-codegen.h"
