+++ gcc/d/d-incpath.cc
@@ -28,6 +28,7 @@
 #include "options.h"
 #include "cppdefault.h"
 
+#include "d-tree.h"
 #include "d-lang.h"
 #include "d-codegen.h"
