+++ gcc/d/d-lang.cc
@@ -51,6 +51,7 @@
 #include "hash-set.h"
 #include "function.h"
 
+#include "d-tree.h"
 #include "d-lang.h"
 #include "d-codegen.h"
 #include "d-objfile.h"
