+++ gcc/d/d-longdouble.cc
@@ -28,7 +28,7 @@
 #include "diagnostic.h"
 #include "stor-layout.h"
 
-#include "d-lang.h"
+#include "d-tree.h"
 #include "d-codegen.h"
 #include "longdouble.h"
