+++ gcc/d/types.cc
@@ -37,6 +37,7 @@
 #include "stor-layout.h"
 #include "attribs.h"
 
+#include "d-tree.h"
 #include "d-lang.h"
 #include "d-codegen.h"
 #include "d-objfile.h"
